Move the location of blobstorage file in root dir
